requirements.txt
Flask
pandas
gspread
google-auth
qrcode
openpyxl
html5-qrcode==0.0.0   # placeholder; actual js lib يحمّل من CDN في القالب

